Itrack_fastapi_server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-dotenv>=1.0.0
openai>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.35.0
typing-extensions>=4.8.0